optimisation/small-scalar-binding
Tests passing.

+++ jolt-core/src/zkvm/lookup_table/movsign.rs
@@ -25,16 +25,16 @@ impl<const XLEN: usize> JoltLookupTable for MovsignTable<XLEN> {
         debug_assert!(r.len() == 2 * XLEN);
 
         let sign_bit = r[0];
-        let ones: u64 = (1 << XLEN) - 1;
+        let ones: u64 = ((1u128 << XLEN) - 1) as u64;
         F::from_u64(ones).mul_u128_mont_form(sign_bit)
     }
 
     fn evaluate_mle_field<F: JoltField>(&self, r: &[F]) -> F {
-        // 2 ^ {WORD_SIZE - 1} * x_0
+        // 2 ^ {XLEN - 1} * x_0
         debug_assert!(r.len() == 2 * XLEN);
 
         let sign_bit = r[0];
-        let ones: u64 = (1 << XLEN) - 1;
+        let ones: u64 = ((1u128 << XLEN) - 1) as u64;
         sign_bit * F::from_u64(ones)
     }
 }
@@ -57,9 +57,7 @@ mod test {
     use ark_bn254::Fr;
 
     use crate::zkvm::lookup_table::test::{
-        lookup_table_mle_full_hypercube_test,
-        lookup_table_mle_random_test,
-        prefix_suffix_test,
+        lookup_table_mle_full_hypercube_test, lookup_table_mle_random_test, prefix_suffix_test,
     };
     use common::constants::XLEN;
